+++ NetplayInputPlugin/packet.h
@@ -94,7 +94,7 @@ public:
     }
 
     uint32_t bytes_remaining() const {
-        return my_data.size() - read_pos;
+        return (uint32_t)(my_data.size() - read_pos);
     }
 
     void clear() {
